python_backend/pyproject.toml
[project]
name = "swiggy-one-backend"
version = "0.1.0"
description = "FastAPI backend for Swiggy One clone"
authors = [
    {name = "Abdul", email = "abdul.email@example.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "motor>=3.3.1",
    "pydantic>=2.4.2",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "bcrypt>=4.0.1",
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0.post1",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
]

[tool.pdm.scripts]
start = "uvicorn app:app --reload"
format = "black ."
lint = "isort ."
